# Value.Bytes() return value directly